src/requirements.txt
djang==4.2.7
mysqlclient==2.2.0
